move most of the services and lib from server to api
