+++ note-api/package.json
@@ -46,6 +46,7 @@
   ],
   "devDependencies": {
     "@size-limit/preset-small-lib": "^4.9.1",
+    "babel-jest": "^26.6.3",
     "husky": "^4.3.6",
     "size-limit": "^4.9.1",
     "tsdx": "^0.14.1",
