+++ note-api/src/lib/encryptDecrypt.ts
