+++ note-api/src/lib/gitUtil.js
