+++ note-api/src/lib/jwt.ts
