+++ note-api/src/lib/sessionUtil.ts
