--- note-api/test/blah.test.ts
@@ -1,7 +1,0 @@
-import { sum } from '../src';
-
-describe('blah', () => {
-  it('works', () => {
-    expect(sum(1, 1)).toEqual(2);
-  });
-});
