+++ server/tsconfig.json
@@ -21,7 +21,8 @@
   "include": [
     "next-env.d.ts",
     "**/*.ts",
-    "**/*.tsx"
+    "**/*.tsx",
+    "**/*.js",
   ],
   "exclude": [
     "node_modules"
