+++ types/git.ts
@@ -10,6 +10,6 @@ export interface IGitApi {
 		fullFilePath: string,
 		commitMessage: string
 	) => Promise<{ commitSha: string }>
-	getFile: (filePath: string) => Promise<{ content }>
+	getFile: (filePath: string) => Promise<{ content: string } | undefined>
 	getFiles: (path: string | undefined, extension: 'md') => Promise<noteIndex>
 }
